fix: handle Scala 3 nightly version for community build
This was originally reported on Discord but it looks like the
"unmanaged" community build is filtering out Mill projects that use
mill-tpolecat, as you can see in
https://github.com/VirtusLab/community-build3/blob/927f3751011d120c52a5217dfe8ccd3321f49c42/env/prod/config/filtered-projects.txt#L66-L69.
From the comment and from playing around it looks to be because if you
have a NIGHTLY or RC or anything it will puke. This adds a fallback for
Scala 3 projects to revert back to the last stable release in those
situations. Hopefully this will allow for those listed Mill projects to
still be tested in the community build.